implement some function

+++ Yandere.Output/Services/ImageMarkService.cs
@@ -1,45 +1,93 @@
 ﻿using CsvHelper;
+using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Windows.Forms;
+using Yandere.Output.Models;
 
 namespace Yandere.Output.Services
 {
-    public class ImageMarkService
+    public class ImageMarkService : IDisposable
     {
         public ImageMarkService()
         {
         }
 
-        public async Task AddMarks(IEnumerable<string> idList)
+        public async Task<bool> AddMarks(IEnumerable<int> idList)
         {
             if (!idList.Any())
             {
-                return;
+                return true;
             }
-            //TODO
-            var existFile = new StreamReader("UserData.csv", true);
-            var existRead = (await existFile.ReadLineAsync()) ?? "";
-            existFile.Close();
-            existFile.Dispose();
-            var existList = existRead?.Split(",").ToList();
-            var addList = new List<string>();
-            foreach (var id in idList)
+
+            var existList = new List<int>();
+            var exist = File.Exists("UserData.csv");
+            if (exist)
             {
-                if (!existList.Contains(id))
+                var existFile = new StreamReader("UserData.csv", true);
+                var existRead = new CsvReader(existFile, CultureInfo.InvariantCulture);
+                try
                 {
-                    addList.Add(id);
+                    existList = existRead.GetRecords<MarkInfo>().Select(x => x.id).ToList();
+                    existRead.Dispose();
+                    existFile.Close();
+                    existFile.Dispose();
+                }
+                catch (Exception e)
+                {
+                    existRead.Dispose();
+                    existFile.Close();
+                    existFile.Dispose();
+                    throw e;
                 }
             }
-            existList.AddRange(idList);
+            var alert = MessageBox.Show("Unknown error occurred from your manifest,\r\n could you want to delete old file?", "Error", MessageBoxButtons.YesNo);
+            if (alert == DialogResult.No)
+            {
+                return false;
+            }
 
+            var addList = new List<MarkInfo>();
             var text = new StreamWriter("UserData.csv", true);
             var helper = new CsvWriter(text, CultureInfo.InvariantCulture);
-            helper.WriteField(addList);
-            await helper.FlushAsync();
-            await helper.DisposeAsync();
+            try
+            {
+                if (!exist)
+                {
+                    helper.WriteHeader(typeof(MarkInfo));
+                    helper.NextRecord();
+                }
+                foreach (var id in idList)
+                {
+                    if (!existList.Contains(id))
+                    {
+                        helper.WriteRecord(new MarkInfo() { id = id, Saved = false });
+                        helper.NextRecord();
+                    }
+                }
+
+                text.Close();
+                text.Dispose();
+                await helper.FlushAsync();
+                await helper.DisposeAsync();
+            }
+            catch (Exception e)
+            {
+                text.Close();
+                text.Dispose();
+                await helper.FlushAsync();
+                await helper.DisposeAsync();
+                throw e;
+            }
+            return true;
+        }
+
+        public void Dispose()
+        {
+
         }
     }
 }
